Fix package grouping of tests
