--- lib/src/test/java/net/opatry/todoist/entity/TodoistServiceLabelTest.kt
@@ -1,284 +1,0 @@
-/*
- * Copyright (c) 2023 Olivier Patry
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
- * CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
- * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
- * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package net.opatry.todoist.entity
-
-import io.ktor.client.engine.mock.MockEngine
-import io.ktor.client.engine.mock.respond
-import io.ktor.client.plugins.ClientRequestException
-import io.ktor.client.request.HttpRequestData
-import io.ktor.http.HttpHeaders
-import io.ktor.http.HttpMethod
-import io.ktor.http.HttpStatusCode
-import io.ktor.http.headersOf
-import io.ktor.utils.io.ByteReadChannel
-import kotlinx.coroutines.runBlocking
-import net.opatry.todoist.entity.data.labelData
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertThrows
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.JUnit4
-
-@RunWith(JUnit4::class)
-class TodoistServiceLabelTest {
-
-    @Test
-    fun `TodoistService getLabels`() {
-        val testData = labelData.first()
-
-        var request: HttpRequestData? = null
-        val mockEngine = MockEngine {
-            request = it
-            respond(
-                content = ByteReadChannel("[${testData.jsonPayload}]"),
-                status = HttpStatusCode.OK,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            val labels = todoService.getLabels()
-            assertEquals("/v2/labels", request?.url?.encodedPath)
-            assertEquals("", request?.url?.encodedQuery)
-            assertEquals(HttpMethod.Get, request?.method)
-            assertEquals(listOf(testData.expectedEntity), labels)
-        }
-    }
-
-    @Test
-    fun `TodoistService getLabels failure`() {
-        val mockEngine = MockEngine {
-            respond(
-                content = ByteReadChannel(""),
-                status = HttpStatusCode.Forbidden,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            assertThrows(ClientRequestException::class.java) {
-                runBlocking {
-                    todoService.getLabels()
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `TodoistService createLabel`() {
-        val testData = labelData.first()
-
-        var request: HttpRequestData? = null
-        val mockEngine = MockEngine {
-            request = it
-            respond(
-                content = ByteReadChannel(
-                    """{
-                        "id": "2156154810",
-                        "name": "Food",
-                        "color": "charcoal",
-                        "order": 1,
-                        "is_favorite": false
-                    }""".trimIndent()
-                ),
-                status = HttpStatusCode.OK,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            val label = todoService.createLabel(TodoistLabelCreationRequest("Foo"))
-            assertEquals("/v2/labels", request?.url?.encodedPath)
-            assertEquals("", request?.url?.encodedQuery)
-            assertEquals(HttpMethod.Post, request?.method)
-            assertEquals(testData.expectedEntity, label)
-        }
-    }
-
-    @Test
-    fun `TodoistService createLabel failure`() {
-        val mockEngine = MockEngine {
-            respond(
-                content = ByteReadChannel(""),
-                status = HttpStatusCode.Forbidden,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            assertThrows(ClientRequestException::class.java) {
-                runBlocking {
-                    todoService.createLabel(TodoistLabelCreationRequest("foo"))
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `TodoistService getLabel`() {
-        val testData = labelData.first()
-
-        var request: HttpRequestData? = null
-        val mockEngine = MockEngine {
-            request = it
-            respond(
-                content = ByteReadChannel(
-                    """{
-                        "id": "2156154810",
-                        "name": "Food",
-                        "color": "charcoal",
-                        "order": 1,
-                        "is_favorite": false
-                    }""".trimIndent()
-                ),
-                status = HttpStatusCode.OK,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            val label = todoService.getLabel("2156154810")
-            assertEquals("/v2/labels/2156154810", request?.url?.encodedPath)
-            assertEquals("", request?.url?.encodedQuery)
-            assertEquals(HttpMethod.Get, request?.method)
-            assertEquals(testData.expectedEntity, label)
-        }
-    }
-
-    @Test
-    fun `TodoistService getLabel failure`() {
-        val mockEngine = MockEngine {
-            respond(
-                content = ByteReadChannel(""),
-                status = HttpStatusCode.Forbidden,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            assertThrows(ClientRequestException::class.java) {
-                runBlocking {
-                    todoService.getLabel("42")
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `TodoistService deleteLabel`() {
-        var request: HttpRequestData? = null
-        val mockEngine = MockEngine {
-            request = it
-            respond(
-                content = ByteReadChannel(
-                    """{
-                        "id": "2156154810",
-                        "name": "Drinks",
-                        "color": "charcoal",
-                        "order": 1,
-                        "is_favorite": false
-                    }""".trimIndent()
-                ),
-                status = HttpStatusCode.NoContent,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            todoService.deleteLabel("2156154810")
-            assertEquals("/v2/labels/2156154810", request?.url?.encodedPath)
-            assertEquals("", request?.url?.encodedQuery)
-            assertEquals(HttpMethod.Delete, request?.method)
-        }
-    }
-
-    @Test
-    fun `TodoistService updateLabel failure`() {
-        val mockEngine = MockEngine {
-            respond(
-                content = ByteReadChannel(""),
-                status = HttpStatusCode.Forbidden,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            assertThrows(ClientRequestException::class.java) {
-                runBlocking {
-                    todoService.updateLabel("42", TodoistLabelUpdateRequest("toto"))
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `TodoistService updateLabel`() {
-        val testData = labelData.first()
-
-        var request: HttpRequestData? = null
-        val mockEngine = MockEngine {
-            request = it
-            respond(
-                content = ByteReadChannel(
-                    """{
-                        "id": "2156154810",
-                        "name": "Drinks",
-                        "color": "charcoal",
-                        "order": 1,
-                        "is_favorite": false
-                    }""".trimIndent()
-                ),
-                status = HttpStatusCode.OK,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            val label = todoService.updateLabel("2156154810", TodoistLabelUpdateRequest("Drinks"))
-            assertEquals("/v2/labels/2156154810", request?.url?.encodedPath)
-            assertEquals("", request?.url?.encodedQuery)
-            assertEquals(HttpMethod.Post, request?.method)
-            val expected = (testData.expectedEntity as TodoistLabel).copy(name = "Drinks")
-            assertEquals(expected, label)
-        }
-    }
-
-    @Test
-    fun `TodoistService deleteLabel failure`() {
-        val mockEngine = MockEngine {
-            respond(
-                content = ByteReadChannel(""),
-                status = HttpStatusCode.Forbidden,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            assertThrows(ClientRequestException::class.java) {
-                runBlocking {
-                    todoService.deleteLabel("42")
-                }
-            }
-        }
-    }
-}
