+++ lib/src/test/java/net/opatry/todoist/service/TodoistServiceProjectTest.kt
@@ -20,7 +20,7 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.opatry.todoist.entity
+package net.opatry.todoist.service
 
 import io.ktor.client.engine.mock.MockEngine
 import io.ktor.client.engine.mock.respond
@@ -32,7 +32,10 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.http.headersOf
 import io.ktor.utils.io.ByteReadChannel
 import kotlinx.coroutines.runBlocking
-import net.opatry.todoist.entity.data.sectionData
+import net.opatry.todoist.entity.TodoistProject
+import net.opatry.todoist.entity.TodoistProjectCreationRequest
+import net.opatry.todoist.entity.TodoistProjectUpdateRequest
+import net.opatry.todoist.entity.data.projectData
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertThrows
 import org.junit.Test
@@ -40,11 +43,11 @@ import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 
 @RunWith(JUnit4::class)
-class TodoistServiceSectionTest {
+class TodoistServiceProjectTest {
 
     @Test
-    fun `TodoistService getSections`() {
-        val testData = sectionData.first()
+    fun `TodoistService getProjects`() {
+        val testData = projectData.first()
 
         var request: HttpRequestData? = null
         val mockEngine = MockEngine {
@@ -57,16 +60,16 @@ class TodoistServiceSectionTest {
         }
 
         usingTodoistService(mockEngine) { todoService ->
-            val sections = todoService.getSections("42")
-            assertEquals("/v2/sections", request?.url?.encodedPath)
-            assertEquals("project_id=42", request?.url?.encodedQuery)
+            val projects = todoService.getProjects()
+            assertEquals("/v2/projects", request?.url?.encodedPath)
+            assertEquals("", request?.url?.encodedQuery)
             assertEquals(HttpMethod.Get, request?.method)
-            assertEquals(listOf(testData.expectedEntity), sections)
+            assertEquals(listOf(testData.expectedEntity), projects)
         }
     }
 
     @Test
-    fun `TodoistService getSections failure`() {
+    fun `TodoistService getProjects failure`() {
         val mockEngine = MockEngine {
             respond(
                 content = ByteReadChannel(""),
@@ -78,24 +81,32 @@ class TodoistServiceSectionTest {
         usingTodoistService(mockEngine) { todoService ->
             assertThrows(ClientRequestException::class.java) {
                 runBlocking {
-                    todoService.getSections("7025")
+                    todoService.getProjects()
                 }
             }
         }
     }
 
     @Test
-    fun `TodoistService createSection`() {
+    fun `TodoistService createProject`() {
         var request: HttpRequestData? = null
         val mockEngine = MockEngine {
             request = it
             respond(
                 content = ByteReadChannel(
                     """{
-                        "id": "7025",
-                        "project_id": "2203306141",
+                        "id": "2203306141",
+                        "name": "Shopping List",
+                        "comment_count": 0,
+                        "color": "charcoal",
+                        "is_shared": false,
                         "order": 1,
-                        "name": "Groceries"
+                        "is_favorite": true,
+                        "is_inbox_project": false,
+                        "is_team_inbox": false,
+                        "view_style": "list",
+                        "url": "https://todoist.com/showProject?id=2203306141",
+                        "parent_id": null
                     }""".trimIndent()
                 ),
                 status = HttpStatusCode.OK,
@@ -104,60 +115,32 @@ class TodoistServiceSectionTest {
         }
 
         usingTodoistService(mockEngine) { todoService ->
-            val sectionData = TodoistSectionCreationRequest("Groceries", "2203306141")
+            val projectData = TodoistProjectCreationRequest("Shopping List")
 
-            val section = todoService.createSection(sectionData)
-            assertEquals("/v2/sections", request?.url?.encodedPath)
+            val project = todoService.createProject(projectData)
+            assertEquals("/v2/projects", request?.url?.encodedPath)
             assertEquals("", request?.url?.encodedQuery)
             assertEquals(HttpMethod.Post, request?.method)
-            val expected = TodoistSection(
-                id = "7025",
-                projectId = "2203306141",
+            val expected = TodoistProject(
+                id = "2203306141",
+                name = "Shopping List",
+                commentCount = 0,
+                color = "charcoal",
+                isShared = false,
                 order = 1,
-                name = "Groceries"
+                isFavorite = true,
+                isInboxProject = false,
+                isTeamInbox = false,
+                viewStyle = TodoistProject.ViewStyle.List,
+                url = "https://todoist.com/showProject?id=2203306141",
+                parentId = null
             )
-            assertEquals(expected, section)
+            assertEquals(expected, project)
         }
     }
 
     @Test
-    fun `TodoistService createSection with order`() {
-        var request: HttpRequestData? = null
-        val mockEngine = MockEngine {
-            request = it
-            respond(
-                content = ByteReadChannel(
-                    """{
-                        "id": "7025",
-                        "project_id": "2203306141",
-                        "order": 5,
-                        "name": "Groceries"
-                    }""".trimIndent()
-                ),
-                status = HttpStatusCode.OK,
-                headers = headersOf(HttpHeaders.ContentType, "application/json")
-            )
-        }
-
-        usingTodoistService(mockEngine) { todoService ->
-            val sectionData = TodoistSectionCreationRequest("Groceries", "2203306141", 5)
-
-            val section = todoService.createSection(sectionData)
-            assertEquals("/v2/sections", request?.url?.encodedPath)
-            assertEquals("", request?.url?.encodedQuery)
-            assertEquals(HttpMethod.Post, request?.method)
-            val expected = TodoistSection(
-                id = "7025",
-                projectId = "2203306141",
-                order = 5,
-                name = "Groceries"
-            )
-            assertEquals(expected, section)
-        }
-    }
-
-    @Test
-    fun `TodoistService createSection failure`() {
+    fun `TodoistService createProject failure`() {
         val mockEngine = MockEngine {
             respond(
                 content = ByteReadChannel(""),
@@ -169,15 +152,15 @@ class TodoistServiceSectionTest {
         usingTodoistService(mockEngine) { todoService ->
             assertThrows(ClientRequestException::class.java) {
                 runBlocking {
-                    todoService.createSection(TodoistSectionCreationRequest("Foo", "7025"))
+                    todoService.createProject(TodoistProjectCreationRequest("Foo"))
                 }
             }
         }
     }
 
     @Test
-    fun `TodoistService getSection`() {
-        val testData = sectionData.first()
+    fun `TodoistService getProject`() {
+        val testData = projectData.first()
 
         var request: HttpRequestData? = null
         val mockEngine = MockEngine {
@@ -190,16 +173,16 @@ class TodoistServiceSectionTest {
         }
 
         usingTodoistService(mockEngine) { todoService ->
-            val section = todoService.getSection("7025")
-            assertEquals("/v2/sections/7025", request?.url?.encodedPath)
+            val project = todoService.getProject("2203306141")
+            assertEquals("/v2/projects/2203306141", request?.url?.encodedPath)
             assertEquals("", request?.url?.encodedQuery)
             assertEquals(HttpMethod.Get, request?.method)
-            assertEquals(testData.expectedEntity, section)
+            assertEquals(testData.expectedEntity, project)
         }
     }
 
     @Test
-    fun `TodoistService getSection failure`() {
+    fun `TodoistService getProject failure`() {
         val mockEngine = MockEngine {
             respond(
                 content = ByteReadChannel(""),
@@ -211,26 +194,32 @@ class TodoistServiceSectionTest {
         usingTodoistService(mockEngine) { todoService ->
             assertThrows(ClientRequestException::class.java) {
                 runBlocking {
-                    todoService.getSection("7025")
+                    todoService.getProject("2203306141")
                 }
             }
         }
     }
 
     @Test
-    fun `TodoistService updateSection`() {
-        val testData = sectionData.first()
-
+    fun `TodoistService updateProject`() {
         var request: HttpRequestData? = null
         val mockEngine = MockEngine {
             request = it
             respond(
                 content = ByteReadChannel(
                     """{
-                        "id": "7025",
-                        "project_id": "2203306141",
+                        "id": "2203306141",
+                        "name": "Things to buy",
+                        "comment_count": 0,
+                        "color": "charcoal",
+                        "is_shared": false,
                         "order": 1,
-                        "name": "Supermarket"
+                        "is_favorite": false,
+                        "is_inbox_project": false,
+                        "is_team_inbox": false,
+                        "view_style": "list",
+                        "url": "https://todoist.com/showProject?id=2203306141",
+                        "parent_id": null
                     }""".trimIndent()
                 ),
                 status = HttpStatusCode.OK,
@@ -239,17 +228,30 @@ class TodoistServiceSectionTest {
         }
 
         usingTodoistService(mockEngine) { todoService ->
-            val section = todoService.updateSection("7025", TodoistSectionUpdateRequest(name = "Supermarket"))
-            assertEquals("/v2/sections/7025", request?.url?.encodedPath)
+            val project = todoService.updateProject("2203306141", TodoistProjectUpdateRequest(name = "Things to buy"))
+            assertEquals("/v2/projects/2203306141", request?.url?.encodedPath)
             assertEquals("", request?.url?.encodedQuery)
             assertEquals(HttpMethod.Post, request?.method)
-            val expected = (testData.expectedEntity as TodoistSection).copy(name = "Supermarket")
-            assertEquals(expected, section)
+            val expected = TodoistProject(
+                id = "2203306141",
+                name = "Things to buy",
+                commentCount = 0,
+                color = "charcoal",
+                isShared = false,
+                order = 1,
+                isFavorite = false,
+                isInboxProject = false,
+                isTeamInbox = false,
+                viewStyle = TodoistProject.ViewStyle.List,
+                url = "https://todoist.com/showProject?id=2203306141",
+                parentId = null
+            )
+            assertEquals(expected, project)
         }
     }
 
     @Test
-    fun `TodoistService updateSection failure`() {
+    fun `TodoistService updateProject failure`() {
         val mockEngine = MockEngine {
             respond(
                 content = ByteReadChannel(""),
@@ -261,14 +263,14 @@ class TodoistServiceSectionTest {
         usingTodoistService(mockEngine) { todoService ->
             assertThrows(ClientRequestException::class.java) {
                 runBlocking {
-                    todoService.updateSection("7025", TodoistSectionUpdateRequest(name = "Supermarket"))
+                    todoService.updateProject("2995104339", TodoistProjectUpdateRequest(name = "Bar"))
                 }
             }
         }
     }
 
     @Test
-    fun `TodoistService deleteSection`() {
+    fun `TodoistService deleteProject`() {
         var request: HttpRequestData? = null
         val mockEngine = MockEngine {
             request = it
@@ -280,15 +282,15 @@ class TodoistServiceSectionTest {
         }
 
         usingTodoistService(mockEngine) { todoService ->
-            todoService.deleteSection("2992679862")
-            assertEquals("/v2/sections/2992679862", request?.url?.encodedPath)
+            todoService.deleteProject("2992679862")
+            assertEquals("/v2/projects/2992679862", request?.url?.encodedPath)
             assertEquals("", request?.url?.encodedQuery)
             assertEquals(HttpMethod.Delete, request?.method)
         }
     }
 
     @Test
-    fun `TodoistService deleteSection failure`() {
+    fun `TodoistService deleteProject failure`() {
         val mockEngine = MockEngine {
             respond(
                 content = ByteReadChannel(""),
@@ -300,7 +302,7 @@ class TodoistServiceSectionTest {
         usingTodoistService(mockEngine) { todoService ->
             assertThrows(ClientRequestException::class.java) {
                 runBlocking {
-                    todoService.deleteSection("2992679862")
+                    todoService.deleteProject("2992679862")
                 }
             }
         }
